creating tests for the register page
